spark_etl/requirements.txt
pyspark==3.5.0
boto3==1.34.0
pandas==2.0.3
pyarrow==12.0.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2

